change to protocol: OHAI includes client timestamp
This fixes a rather serious bug that occurs when the controller's clock is
not synchronized with the server clock. This can happen with surprising
frequency with beaglebones, and it leads to garbage messages in the
logs. Now, the controller has to send its current timestamp and the host
will reject it if it isn't within 10 s